src/test/java/hellospring/springstudy/service/MemberServiceTest.java
package hellospring.springstudy.service;

import hellospring.springstudy.domain.Member;
import hellospring.springstudy.domain.repository.MemoryMemberRepository;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.util.Optional;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.junit.jupiter.api.Assertions.*;

// 테스트는 빌드할 때 실제 코드에 포함되지 않음.
class MemberServiceTest {

    MemberService memberService;
    MemoryMemberRepository memberRepository;

    // DI : Dependency Injection
    @BeforeEach
    public void beforeEach() {
        memberRepository = new MemoryMemberRepository();
        memberService = new MemberService(memberRepository);
    }

    @AfterEach
    public void afterEach() {
        memberRepository.clearStore();
    }

    @Test // 테스트는 메서드를 한글로 작성 가능
    void 회원가입() {
        // given (값이 주어졌을 때)
        Member member = new Member();
        member.setName("hello");

        // when (어떻게 실행을 하며)
        Long saveId = memberService.join(member);

        // then (어떤 결과를 얻을 수 있는가)
        Member findMember = memberService.findOne(saveId).get();
        assertThat(member.getName()).isEqualTo(findMember.getName());
    }

    @Test
    public void 중복_회원_예외() {
        // given
        Member member1 = new Member();
        member1.setName("spring");

        Member member2 = new Member();
        member2.setName("spring");

        // when
        memberService.join(member1);
        IllegalStateException e = assertThrows(IllegalStateException.class, () -> memberService.join(member2));

        assertThat(e.getMessage()).isEqualTo("이미 존재하는 회원입니다.");

//        try {
//            memberService.join(member2);
//            fail("예외가 발생해야 합니다.");
//        } catch (IllegalStateException e) {
//            assertThat(e.getMessage()).isEqualTo("이미 존재하는 회원입니다.");
//        }

        // then
    }

    @Test
    void findMembers() {
    }

    @Test
    void findOne() {
    }
}